requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-redis==6.0.0
djangorestframework==3.16.1
idna==3.11
packaging==25.0
pillow==12.1.0
python-dotenv==1.1.1
redis==7.1.0
requests==2.32.5
setuptools==80.10.1
sqlparse==0.5.5
urllib3==2.6.3
wheel==0.46.3
